pyproject.toml
[tool.poetry]
name = "d4ac"
version = "0.1.0"
description = ""
authors = ["Mikio Nakano <mikio.nakano@c4a.jp>"]
packages = [
    { include = "src/*.py"},
    { include = "src/d4ac_main/*.py"},
   { include = "src/facepp/*.py" },
    { include = "src/dialog_server/*.py" },
    {include = "src/mediapipe/*.py"},
    {include = "src/util/*.py"},
    {include = "src/video_input/*.py" },
    {include = "src/user_status/*.py" }
]
exclude =["src/android/*", "src/frontend/*" ]
include = ["src/d4ac_main/static/css/*.css",
"src/config-template.yml",
"src/d4ac_main/static/images/*" ,
"src/d4ac_main/static/js/*" ,
"src/d4ac_main/templates/*" ]
 

[tool.poetry.dependencies]
python = "^3.8"
aiofiles="^0.6.0"
altgraph="^0.17"
certifi="^2020.12.5"
click="^7.1.2"
fastapi="^0.63.0"
future="^0.18.2"
h11="^0.12.0"
idna="^2.10"
importlib-metadata="^3.9.0"
imutils="^0.5.4"
Jinja2="^2.11.3"
MarkupSafe="^1.1.1"
numpy="^1.20.2"
opencv-python="^4.5.1.48"
paho-mqtt="^1.5.1"
pefile="2019.4.18"
pip-licenses="^3.3.1"
PTable="^0.9.2"
pydantic="^1.8.1"
python-aiml="^0.9.3"
pywin32-ctypes="^0.2.0"
PyYAML="^6.0"
starlette="^0.13.6"
typing-extensions="^3.7.4.3"
urllib3="^1.26.4"
uvicorn="^0.13.4"
zipp="^3.4.1"
pandas="^1.4.1"
openpyxl="^3.0.9"
PySimpleGUI = "^4.60.3"
requests = "^2.28.1"
azure-cognitiveservices-vision-face = "^0.6.0"
mediapipe =  { version = "^0.9.0" ,markers = "sys_platform == 'win32'" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
d4ac = 'src.d4ac:main'
